datasize: split alloc from std
