+++ datasize/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "datasize"
 version = "0.2.15"
-authors = [ "Marc Brinkmann <marc@casperlabs.io>" ]
+authors = ["Marc Brinkmann <marc@casperlabs.io>"]
 edition = "2018"
 license = "MIT OR Apache-2.0"
 description = "A simplified heap memory size estimator"
@@ -9,24 +9,25 @@ documentation = "https://docs.rs/datasize"
 repository = "https://github.com/casperlabs/datasize-rs"
 
 [features]
-default = [ "std", "const-generics" ]
-detailed = [ "std", "serde", "datasize_derive/detailed" ]
-fake_clock-types = [ "fake_instant" ]
-futures-types = [ "futures" ]
-smallvec-types = [ "smallvec", "std" ]
-std = []
-tokio-types = [ "tokio" ]
+default = ["std", "const-generics"]
+detailed = ["std", "serde", "datasize_derive/detailed"]
+fake_clock-types = ["fake_instant"]
+futures-types = ["futures"]
+smallvec-types = ["smallvec", "std"]
+std = ["alloc"]
+alloc = []
+tokio-types = ["tokio", "std"]
 const-generics = []
 
 [dependencies]
 datasize_derive = { version = "0.2.15" }
 fake_instant = { version = "0.4.0", optional = true }
 futures = { version = "0.3.5", optional = true }
-serde = { version = "1", optional = true, features = [ "derive" ] }
+serde = { version = "1", optional = true, features = ["derive"] }
 smallvec = { version = "1.4.2", optional = true }
 tokio = { version = "0.2.22", default-features = false, optional = true, features = [
-  "rt-core",
-  "sync",
+    "rt-core",
+    "sync",
 ] }
 
 [package.metadata.docs.rs]
