+++ datasize/src/fake_clock.rs
@@ -1,3 +1,3 @@
-use super::{non_dynamic_const_heap_size, DataSize};
+use crate::non_dynamic_const_heap_size;
 
 non_dynamic_const_heap_size!(fake_instant::FakeClock, 0);
